verbs: Make loading the verbs_sysfs_dev more robust
Holding a directory FD open across multiple reads to sysfs files
in the same directory prevents the kernel from removing the directory
and replacing it with one that has the same name.

This makes the loading of the sysfs information much more atomic
in the face of device rename/hotplug/etc.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>