azp: Set up CI with Azure Pipelines
Azure Pipelines is intended to replace travis as the CI for rdma-core.
Build times are much faster at around 5 mins (vs 23 mins for travis) and
because it is pre-built container based it should not suffer from the same
flakiness that travis historically has.

This commit replicates all the functionality of travis, except for the
.tar.gz upload on release. That requires a more careful security analysis.

The GitHub experience is similar with the notable difference that
checkpatch warnings/errors now result in a green checkmark. Clicking
through the check will show a count of errors/warnings and Azure's GUI
will show an orange checkmark.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>